fix: move appeal specific code out of controller shared with lpaq
